refactor: 커스텀 예외에 상속 적용
