+++ src/main/java/roomescape/exception/CustomNotFoundException.java
@@ -1,15 +1,8 @@
 package roomescape.exception;
 
-public class CustomNotFoundException extends RuntimeException {
-
-    private final ErrorCode errorCode;
+public class CustomNotFoundException extends CustomException {
 
     public CustomNotFoundException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
+        super(errorCode);
     }
 }
