+++ src/main/java/roomescape/exception/CustomUnprocessableEntityException.java
@@ -1,15 +1,8 @@
 package roomescape.exception;
 
-public class CustomUnprocessableEntityException extends RuntimeException {
-
-    private final ErrorCode errorCode;
+public class CustomUnprocessableEntityException extends CustomException {
 
     public CustomUnprocessableEntityException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
+        super(errorCode);
     }
 }
